Move goal where clause building to new WhereBuilder module
